Fix bytea in Image

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -1,46 +1,46 @@
 create table comments
 (
-    id       int8         not null,
+    id       serial       not null,
     created  timestamp(6),
     message  text         not null,
     username varchar(255) not null,
-    user_id  int8         not null,
-    post_id  int8,
+    user_id  serial       not null,
+    post_id  serial,
     primary key (id)
 );
 create table images
 (
-    id      int8         not null,
+    id      serial       not null,
     image   bytea,
     name    varchar(255) not null,
-    post_id int8,
-    user_id int8,
+    post_id serial,
+    user_id serial,
     primary key (id)
 );
 create table post_liked_users
 (
-    post_id     int8 not null,
+    post_id     serial not null,
     liked_users varchar(255)
 );
 create table posts
 (
-    id       int8 not null,
+    id       serial not null,
     created  timestamp(6),
     likes    integer,
     location varchar(255),
     text     varchar(255),
     title    varchar(255),
-    user_id  int8,
+    user_id  serial,
     primary key (id)
 );
 create table users
 (
-    id        serial     not null,
+    id        serial       not null,
     bio       text,
     created   timestamp(6),
     email     varchar(255),
     firstname varchar(255) not null,
-    lastname varchar(255) not null,
+    lastname  varchar(255) not null,
     username  varchar(255),
     password  varchar(255),
     role      varchar(255),
